Added sending DUC on queue
